requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
werkzeug==3.0.3
requests==2.32.3
gunicorn==22.0.0
psycopg2-binary==2.9.9
stripe==9.12.0
python-slugify==8.0.4
pywebpush==2.0.0
pyotp==2.9.0
qrcode==7.4.2
webauthn==2.7.1
Pillow==10.4.0
pytest==8.2.0
pytest-flask==1.3.0
